refactoring dp review: move entity factories to modules
