+++ apps/server/src/modules/authorization/domain/rules/context-external-tool.rule.spec.ts
@@ -1,18 +1,13 @@
 import { ContextExternalTool } from '@modules/tool/context-external-tool/domain';
 import { ContextExternalToolEntity } from '@modules/tool/context-external-tool/entity';
-import { contextExternalToolEntityFactory } from '@modules/tool/context-external-tool/entity/testing';
+import { contextExternalToolEntityFactory } from '@modules/tool/context-external-tool/testing';
 import { SchoolExternalTool } from '@modules/tool/school-external-tool/domain';
 import { SchoolExternalToolEntity } from '@modules/tool/school-external-tool/entity';
+import { schoolExternalToolEntityFactory } from '@modules/tool/school-external-tool/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Role, User } from '@shared/domain/entity';
 import { Permission } from '@shared/domain/interface';
-import {
-	roleFactory,
-	schoolEntityFactory,
-	schoolExternalToolEntityFactory,
-	setupEntities,
-	userFactory,
-} from '@shared/testing';
+import { roleFactory, schoolEntityFactory, setupEntities, userFactory } from '@shared/testing';
 import { AuthorizationHelper } from '../service/authorization.helper';
 import { Action } from '../type';
 import { ContextExternalToolRule } from './context-external-tool.rule';
