+++ apps/server/src/modules/authorization/domain/rules/school-external-tool.rule.spec.ts
@@ -1,12 +1,12 @@
 import { SchoolExternalTool } from '@modules/tool/school-external-tool/domain';
 import { SchoolExternalToolEntity } from '@modules/tool/school-external-tool/entity';
+import { schoolExternalToolEntityFactory } from '@modules/tool/school-external-tool/testing';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Role, User } from '@shared/domain/entity';
 import { Permission } from '@shared/domain/interface';
 import {
 	roleFactory,
 	schoolEntityFactory,
-	schoolExternalToolEntityFactory,
 	schoolExternalToolFactory,
 	setupEntities,
 	userFactory,
