+++ apps/server/src/modules/board/repo/board-do.repo.spec.ts
@@ -4,8 +4,8 @@ import { NotFoundError } from '@mikro-orm/core';
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { FilesStorageClientAdapterService } from '@modules/files-storage-client';
 import { DrawingElementAdapterService } from '@modules/tldraw-client';
-import { contextExternalToolEntityFactory } from '@modules/tool/context-external-tool/entity/testing';
 import { ContextExternalToolService } from '@modules/tool/context-external-tool/service';
+import { contextExternalToolEntityFactory } from '@modules/tool/context-external-tool/testing';
 import { NotFoundException } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AnyBoardDo, BoardExternalReferenceType, Card, Column, ColumnBoard } from '@shared/domain/domainobject';
