+++ apps/server/src/modules/tool/context-external-tool/controller/api-test/tool-context.api.spec.ts
@@ -7,11 +7,8 @@ import { Permission } from '@shared/domain/interface';
 import {
 	accountFactory,
 	courseFactory,
-	customParameterEntityFactory,
-	externalToolEntityFactory,
 	roleFactory,
 	schoolEntityFactory,
-	schoolExternalToolEntityFactory,
 	TestApiClient,
 	UserAndAccountTestFactory,
 	userFactory,
@@ -19,9 +16,11 @@ import {
 import { ObjectId } from 'bson';
 import { CustomParameterScope, CustomParameterType, ToolContextType } from '../../../common/enum';
 import { ExternalToolEntity } from '../../../external-tool/entity';
+import { customParameterEntityFactory, externalToolEntityFactory } from '../../../external-tool/testing';
 import { SchoolExternalToolEntity } from '../../../school-external-tool/entity';
+import { schoolExternalToolEntityFactory } from '../../../school-external-tool/testing';
 import { ContextExternalToolEntity, ContextExternalToolType } from '../../entity';
-import { contextExternalToolEntityFactory } from '../../entity/testing';
+import { contextExternalToolEntityFactory } from '../../testing';
 import {
 	ContextExternalToolPostParams,
 	ContextExternalToolResponse,
