+++ apps/server/src/modules/tool/context-external-tool/controller/api-test/tool-reference.api.spec.ts
@@ -9,19 +9,19 @@ import {
 	contextExternalToolConfigurationStatusResponseFactory,
 	courseFactory,
 	customParameterFactory,
-	externalToolEntityFactory,
 	schoolEntityFactory,
-	schoolExternalToolEntityFactory,
 	TestApiClient,
 	UserAndAccountTestFactory,
 } from '@shared/testing';
 
 import { Response } from 'supertest';
 import { CustomParameterLocation, CustomParameterScope, ToolContextType } from '../../../common/enum';
-import { contextExternalToolEntityFactory } from '../../entity/testing';
 import { ExternalToolEntity } from '../../../external-tool/entity';
+import { externalToolEntityFactory } from '../../../external-tool/testing';
 import { SchoolExternalToolEntity } from '../../../school-external-tool/entity';
+import { schoolExternalToolEntityFactory } from '../../../school-external-tool/testing';
 import { ContextExternalToolEntity, ContextExternalToolType } from '../../entity';
+import { contextExternalToolEntityFactory } from '../../testing';
 import { ContextExternalToolContextParams, ToolReferenceListResponse, ToolReferenceResponse } from '../dto';
 
 describe('ToolReferenceController (API)', () => {
