--- apps/server/src/modules/tool/context-external-tool/entity/testing/index.ts
@@ -1,1 +1,0 @@
-export { contextExternalToolEntityFactory } from './factory/context-external-tool-entity.factory';
