+++ apps/server/src/modules/tool/context-external-tool/testing/index.ts
@@ -1,0 +1,1 @@
+export { contextExternalToolEntityFactory } from './context-external-tool-entity.factory';
