+++ apps/server/src/modules/tool/external-tool/testing/index.ts
@@ -1,0 +1,1 @@
+export { externalToolEntityFactory, customParameterEntityFactory } from './external-tool-entity.factory';
