+++ apps/server/src/modules/tool/school-external-tool/entity/school-external-tool.entity.spec.ts
@@ -1,14 +1,9 @@
-import {
-	basicToolConfigFactory,
-	customParameterEntityFactory,
-	externalToolEntityFactory,
-	schoolEntityFactory,
-	setupEntities,
-} from '@shared/testing';
+import { basicToolConfigFactory, schoolEntityFactory, setupEntities } from '@shared/testing';
 import { schoolExternalToolConfigurationStatusEntityFactory } from '@shared/testing/factory/school-external-tool-configuration-status-entity.factory';
-import { schoolExternalToolEntityFactory } from '@shared/testing/factory/school-external-tool-entity.factory';
 import { CustomParameterLocation, CustomParameterScope, CustomParameterType, ToolConfigType } from '../../common/enum';
 import { CustomParameterEntity, ExternalToolConfigEntity, ExternalToolEntity } from '../../external-tool/entity';
+import { customParameterEntityFactory, externalToolEntityFactory } from '../../external-tool/testing';
+import { schoolExternalToolEntityFactory } from '../testing';
 import { SchoolExternalToolEntity } from './school-external-tool.entity';
 
 describe('SchoolExternalToolEntity', () => {
