+++ apps/server/src/modules/tool/school-external-tool/testing/index.ts
@@ -1,0 +1,1 @@
+export { schoolExternalToolEntityFactory } from './school-external-tool-entity.factory';
