+++ apps/server/src/modules/tool/school-external-tool/testing/school-external-tool-entity.factory.ts
@@ -1,8 +1,8 @@
+import { externalToolEntityFactory } from '@modules/tool/external-tool/testing';
 import { SchoolExternalToolEntity, SchoolExternalToolProperties } from '@modules/tool/school-external-tool/entity';
 import { BaseFactory } from '@shared/testing/factory/base.factory';
-import { externalToolEntityFactory } from './external-tool-entity.factory';
-import { schoolEntityFactory } from './school-entity.factory';
-import { schoolExternalToolConfigurationStatusEntityFactory } from './school-external-tool-configuration-status-entity.factory';
+import { schoolEntityFactory } from '@shared/testing/factory/school-entity.factory';
+import { schoolExternalToolConfigurationStatusEntityFactory } from '@shared/testing/factory/school-external-tool-configuration-status-entity.factory';
 
 export const schoolExternalToolEntityFactory = BaseFactory.define<
 	SchoolExternalToolEntity,
