+++ apps/server/src/shared/testing/factory/index.ts
@@ -9,7 +9,6 @@ export * from './course.factory';
 export * from './coursegroup.factory';
 export * from './domainobject';
 export * from './external-group-dto.factory';
-export * from './external-tool-entity.factory';
 export * from './external-tool-pseudonym.factory';
 export * from './federal-state.factory';
 export * from './filerecord.factory';
@@ -21,7 +20,6 @@ export * from './material.factory';
 export * from './news.factory';
 export * from './role-dto.factory';
 export * from './role.factory';
-export * from './school-external-tool-entity.factory';
 export * from './school-entity.factory';
 export * from './schoolyear.factory';
 export * from './share-token.do.factory';
